requirements.txt
autogluon
numba==0.56.0
pillow
tensorflow
keras
shutil
random
